+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.8.1] - 2017-03-03
+### Added
+* [schema.org] `image` → blog entries ([content.php])
+* [schema.org] `publisher` → blog entries ([content.php])
+
 ### Changed
-* byline `<time>` [schema.org] `dateCreated` → `datePublished` ([content.php])
-* [schema.org] `dateCreated` → byline `<meta>` ([content.php])
+* blog entries byline `<time>` [schema.org] `dateCreated` → `datePublished` ([content.php])
+* [schema.org] `dateCreated` → blog entries byline `<meta>` ([content.php])
 
 ### Fixed
 * [schema.org] `BlogPosting` errors
@@ -241,7 +247,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * [LISENCE]
 * [README.md]
 
-[Unreleased]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.7.4...v0.8.0
 [0.7.4]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.7.3...v0.7.4
 [0.7.3]: https://github.com/AdamSteffanick/wp2012-s/compare/v0.7.2...v0.7.3